doBusiness implementation

+++ code-challenge/src/main/java/com/statista/code/challenge/controller/BookingDTO.java
@@ -1,5 +1,7 @@
 package com.statista.code.challenge.controller;
 
-public record BookingDTO(String description, Double price, String currency, String subscriptionStartDate, String email,
-                         String departmentName) {
+import java.math.BigDecimal;
+
+public record BookingDTO(String description, BigDecimal price, String currency, String subscriptionStartDate,
+                         String email, String departmentName) {
 }
